Add basic validation to getToken()
When the OAuth2 Authorization Server returns a HTTP 200 OK, the JSON response should still be validated. This protects from the PHP NOTICE errors that are triggered when 'access_token' and/or 'token_type' are missing from $data. By throwing an exception, it becomes clear to the user where things went wrong - especially if the OAuth2 server sets the 'error' and 'error description' properties (like many implementations do).